Edit selected pane

+++ PMDMMO-DataEditor/src/fr/darkxell/dataeditor/application/controls/CutscenesTreeCell.java
@@ -4,37 +4,46 @@ import java.util.Optional;
 
 import com.darkxell.client.mechanics.cutscene.Cutscene;
 
-import fr.darkxell.dataeditor.application.controller.CutscenesTabController;
+import fr.darkxell.dataeditor.application.controller.cutscene.CutscenesTabController;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.ButtonType;
 import javafx.scene.control.ContextMenu;
 import javafx.scene.control.ListCell;
 import javafx.scene.control.MenuItem;
-import javafx.scene.text.Text;
 
 public class CutscenesTreeCell extends ListCell<Cutscene>
 {
 
-	private ContextMenu menu(Cutscene item)
+	private ContextMenu menu(Cutscene item, boolean empty)
 	{
 		ContextMenu menu = new ContextMenu();
 
-		MenuItem edit = new MenuItem("Edit");
-		edit.setOnAction(e -> {
-			CutscenesTabController.instance.onEdit(item);
+		MenuItem add = new MenuItem("New Cutscene...");
+		add.setOnAction(e -> {
+			CutscenesTabController.instance.onCreateCutscene();
 		});
+		menu.getItems().add(add);
 
-		MenuItem remove = new MenuItem("Delete");
-		remove.setOnAction(e -> {
-			Alert alert = new Alert(AlertType.CONFIRMATION);
-			alert.setHeaderText("Delete Cutscene");
-			alert.setContentText(new Text("Are you sure you want to delete Cutscene '" + item.name + "'?").toString());
-			Optional<ButtonType> result = alert.showAndWait();
-			if (result.get() == ButtonType.OK) CutscenesTabController.instance.onDelete(item);
-		});
+		if (!empty)
+		{
+			MenuItem edit = new MenuItem("Edit");
+			edit.setOnAction(e -> {
+				CutscenesTabController.instance.onEdit(item);
+			});
 
-		menu.getItems().add(remove);
+			MenuItem remove = new MenuItem("Delete");
+			remove.setOnAction(e -> {
+				Alert alert = new Alert(AlertType.CONFIRMATION);
+				alert.setHeaderText("Delete Cutscene");
+				alert.setContentText("Are you sure you want to delete Cutscene '" + item.name + "'?");
+				Optional<ButtonType> result = alert.showAndWait();
+				if (result.get() == ButtonType.OK) CutscenesTabController.instance.onDelete(item);
+			});
+
+			menu.getItems().add(edit);
+			menu.getItems().add(remove);
+		}
 
 		return menu;
 	}
@@ -45,7 +54,7 @@ public class CutscenesTreeCell extends ListCell<Cutscene>
 		super.updateItem(item, empty);
 		this.setText(empty ? "" : item.name);
 
-		if (!empty) this.setContextMenu(this.menu(item));
+		this.setContextMenu(this.menu(item, empty));
 	}
 
 }
